Restructure code to support more donation methods
